SPARKC-398: Removed Akka leftovers and refactored build scripts so that the testing configuration is extracted into a separate file
