+++ spark-cassandra-connector/src/test/scala/com/datastax/spark/connector/testkit/package.scala
@@ -1,15 +1,8 @@
 package com.datastax.spark.connector
 
-import scala.collection.immutable
-import scala.concurrent.duration._
-import akka.util.Timeout
-
 package object testkit {
 
   final val DefaultHost = "127.0.0.1"
-
-  implicit val DefaultTimeout = Timeout(5.seconds)
-
 
   val dataSeq = Seq (
     Seq("1first", "1round", "1words"),
@@ -19,7 +12,5 @@ package object testkit {
   )
 
   val data = dataSeq.head
-
-  val actorName = "my-actor"
 
 }
